+++ src/main/java/io/xlorey/fluxloader/shared/ServiceManager.java
@@ -21,8 +21,18 @@ public class ServiceManager {
      * @param serviceInterface service interface
      * @param serviceImplementation service implementation
      */
-    public static <T> void register(Class<T> serviceInterface, T serviceImplementation) {
+    public static synchronized <T> void register(Class<T> serviceInterface, T serviceImplementation) {
         services.put(serviceInterface, serviceImplementation);
+    }
+
+    /**
+     * Removes a service from the list of registered ones
+     * @param serviceInterface service interface
+     * @param <T> service type
+     */
+    public static synchronized <T> void unregister(Class<T> serviceInterface) {
+        if (services.get(serviceInterface) == null) return;
+        services.remove(serviceInterface);
     }
 
     /**
@@ -31,7 +41,7 @@ public class ServiceManager {
      * @param serviceInterface service interface
      * @return an instance of the service corresponding to the specified interface, or {@code null} if the service is not found
      */
-    public static <T> T getService(Class<T> serviceInterface) {
+    public static synchronized <T> T getService(Class<T> serviceInterface) {
         return serviceInterface.cast(services.get(serviceInterface));
     }
 }